+++ CHANGELOG.md
@@ -5,6 +5,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.3.1] — unreleased
 
 -   Make all families fall back to "sans-serif" fonts.
+-   Cache `FontLibrary::face_for_char` glyph lookups.
 
 ## [0.3.0] — 2021-06-15